move to pyproject.toml with hatchling
