+++ tests/test_compare_methods.py
@@ -1,6 +1,5 @@
 from pytest import approx, mark
 import numpy as np
-import sys
 import itertools
 from .conftest import skip_s4_test
 
@@ -1151,7 +1150,7 @@ def test_rcwa_tmm_matrix_profiles(RCWA_method):
 @mark.skipif(skip_s4_test(), reason="Only works if S4 installed")
 @mark.parametrize("RCWA_method", ["S4", "Inkstone"])
 def test_profile_integration(RCWA_method):
-    from rayflare.utilities import get_savepath
+    from rayflare import get_savepath
     import os
     import xarray as xr
